Fixed "sed lstat" error in OS X.
This one is a little tricky to follow. The OS X version of sed understands
"-i" for in-place editing of files; however, the flag is undocumented, and it
apparently must be just before the filename.

Also, I can't seem to manage to get the right quoting syntax to make the sed
insert command work, so I rewrote that part to use a temporary file.

Finally, hid_wrap.c was added as a second target to the SWIG rule, which was
merged with the sed rule.

Tested on 10.3.9 with Python 2.4 and Fink.